+++ src/cli/CLI-Extract.cpp
@@ -86,7 +86,7 @@ CLIExtract::processCmd(const std::filesystem::path& outPath, const LotusLib::Lot
 	// Debug information
 	WarframeExporter::Logger::getInstance().debug("Type Flags: " + std::to_string(types));
 	std::stringstream pkgs;
-	for (int i = 0; i < pkgNames.size(); i++)
+	for (size_t i = 0; i < pkgNames.size(); i++)
 		pkgs << (pkgNames[i]) << " ";
 	WarframeExporter::Logger::getInstance().debug("Loading packages: " + pkgs.str());
